requirements.txt
python-telegram-bot==12.3.0
beautifulsoup4==4.8.2
requests==2.22.0
